Move check form middleware function to new file
Add new file: routes/middleware/commonMiddleware.ts
  - move out checkFormErrors from userRegistrationMiddleware.ts to it
  - update file paths as needed

Add new file: specs/commonMiddleware.spec.ts
  - move out tests for checkFormErrors from userRegistration.spec.ts to it
